+++ pylutron_caseta/pairing.py
@@ -15,7 +15,7 @@ from cryptography.hazmat.primitives import hashes, serialization
 from cryptography.hazmat.primitives.asymmetric import rsa
 from cryptography.x509.oid import NameOID  # type: ignore
 
-from .assets import LAP_CA_PEM, LAP_CERT_PEM, LAP_KEY_PEM
+from .assets import LAP_CA_PEM, LAP_CERT_PEM, LAP_KEY_PEM, LUTRON_ROOT_CA_PEM
 
 LOGGER = logging.getLogger(__name__)
 
@@ -77,9 +77,20 @@ async def async_pair(
         None, _generate_csr_with_ssl_context
     )
 
-    cert_pem, ca_pem = await _async_generate_certificate(
-        server_addr, ssl_context, csr, ready
-    )
+    try:
+        cert_pem, ca_pem = await _async_generate_certificate(
+            server_addr, ssl_context, csr, ready
+        )
+    except ssl.SSLCertVerificationError:
+        # SSL certificate verification error - might be an RA3 processor,
+        # try to connect using the lutron-root certificate instead of LAP_CA
+        ssl_context.load_verify_locations(cadata=LUTRON_ROOT_CA_PEM)
+        cert_pem, ca_pem = await _async_generate_certificate(
+            server_addr, ssl_context, csr, ready
+        )
+        # Generate certificates worked with RA3 lutron-root so bridge is RA3.
+        # Discard the ca_pem for caseta and replace with LUTRON_ROOT_CA_PEM
+        ca_pem = LUTRON_ROOT_CA_PEM
 
     signed_ssl_context = await loop.run_in_executor(
         None, _generate_signed_ssl_context, key_bytes_pem, cert_pem, ca_pem
@@ -117,6 +128,7 @@ async def _async_generate_certificate(
         ),
         timeout=SOCKET_TIMEOUT,
     )
+
     json_socket = JsonSocket(reader, writer)
     LOGGER.info("Press the small black button on the back of the Caseta bridge...")
     if ready is not None: